Adding metaclass
